requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==7.4.3
httpx==0.25.2
cryptography==41.0.7
